+++ src/main/java/com/epam/pharmacy/model/service/impl/DrugPictureServiceImpl.java
@@ -7,20 +7,37 @@ import com.epam.pharmacy.model.dao.EntityTransaction;
 import com.epam.pharmacy.model.entity.DrugPicture;
 import com.epam.pharmacy.model.service.DrugPictureService;
 
+/**
+ * Class-service for working with {@DrugPicture}.
+ * @see DrugPicture
+ * @author Yauheni Tsitou.
+ */
 public class DrugPictureServiceImpl implements DrugPictureService {
+
+    /**
+     * Reference to an object of class {@code DrugPictureServiceImpl}.
+     */
     private static final DrugPictureService instance = new DrugPictureServiceImpl();
 
     private DrugPictureServiceImpl() {
     }
 
+    /**
+     * Method that returns a reference to an object
+     *
+     * @return Reference to an object of class {@code DrugPictureServiceImpl}.
+     */
     public static DrugPictureService getInstance() {
         return instance;
     }
 
+    /**
+     * Reference to an object of class {@code DrugPictureDao}.
+     */
     private static final DrugPictureDao drugPictureDao = DrugPictureDao.getInstance();
 
     @Override
-    public void add(String drugPictureString, int pictureDrugId) throws ServiceException {
+    public void add(String drugPictureString, long pictureDrugId) throws ServiceException {
         if (!drugPictureString.isBlank()) {
             EntityTransaction transaction = new EntityTransaction();
             transaction.init(drugPictureDao);
